+++ src/UnitTestEx.XUnit/Internal/FunctionTesterT.cs
@@ -20,6 +20,7 @@ namespace UnitTestEx.Xunit.Internal
         /// <param name="includeUnitTestConfiguration">Indicates whether to include '<c>appsettings.unittest.json</c>' configuration file.</param>
         /// <param name="includeUserSecrets">Indicates whether to include user secrets.</param>
         /// <param name="additionalConfiguration">Additional configuration values to add/override.</param>
-        internal FunctionTester(ITestOutputHelper output, bool? includeUnitTestConfiguration, bool? includeUserSecrets, IEnumerable<KeyValuePair<string, string>>? additionalConfiguration) : base(new XunitTestImplementor(output), includeUnitTestConfiguration, includeUserSecrets, additionalConfiguration) { }
+        internal FunctionTester(ITestOutputHelper output, bool? includeUnitTestConfiguration, bool? includeUserSecrets, params KeyValuePair<string, string>[] additionalConfiguration) 
+            : base(new XunitTestImplementor(output), includeUnitTestConfiguration, includeUserSecrets, additionalConfiguration) { }
     }
 }