+++ tests/UnitTestEx.Xunit.Test/ServiceBusFunctionTest.cs
@@ -22,7 +22,7 @@ namespace UnitTestEx.Xunit.Test
 
             using var test = CreateFunctionTester<Startup>();
             test.ConfigureServices(sc => mcf.Replace(sc))
-                .Type<ServiceBusFunction>()
+                .ServiceBusTrigger<ServiceBusFunction>()
                 .Run(f => f.Run(new Person { FirstName = "Bob", LastName = "Smith" }, test.Logger))
                 .AssertSuccess();
 
@@ -52,7 +52,7 @@ namespace UnitTestEx.Xunit.Test
 
             using var test = CreateFunctionTester<Startup>();
             test.ConfigureServices(sc => mcf.Replace(sc))
-                .Type<ServiceBusFunction>()
+                .ServiceBusTrigger<ServiceBusFunction>()
                 .Run(f => f.Run(new Person { FirstName = null, LastName = "Smith" }, test.Logger))
                 .AssertException<InvalidOperationException>("First name is required.");
 
@@ -68,7 +68,7 @@ namespace UnitTestEx.Xunit.Test
 
             using var test = CreateFunctionTester<Startup>();
             test.ConfigureServices(sc => mcf.Replace(sc))
-                .Type<ServiceBusFunction>()
+                .ServiceBusTrigger<ServiceBusFunction>()
                 .Run(f => f.Run2(test.CreateServiceBusMessage(new Person { FirstName = "Bob", LastName = "Smith" }), test.Logger))
                 .AssertSuccess();
 
@@ -83,8 +83,8 @@ namespace UnitTestEx.Xunit.Test
                 .Request(HttpMethod.Post, "person").WithJsonBody(new { firstName = "Bob", lastName = (string)null }).Respond.With(HttpStatusCode.InternalServerError);
 
             using var test = CreateFunctionTester<Startup>();
-            var r = test.ConfigureServices(sc => mcf.Replace(sc))
-                .Type<ServiceBusFunction>()
+            test.ConfigureServices(sc => mcf.Replace(sc))
+                .ServiceBusTrigger<ServiceBusFunction>()
                 .Run(f => f.Run2(test.CreateServiceBusMessage(new Person { FirstName = "Bob", LastName = "Smith" }), test.Logger))
                 .AssertException<HttpRequestException>("Response status code does not indicate success: 500 (Internal Server Error).");
 
@@ -98,7 +98,7 @@ namespace UnitTestEx.Xunit.Test
 
             using var test = CreateFunctionTester<Startup>();
             test.ConfigureServices(sc => mcf.Replace(sc))
-                .Type<ServiceBusFunction>()
+                .ServiceBusTrigger<ServiceBusFunction>()
                 .Run(f => f.Run2(test.CreateServiceBusMessage(new Person { FirstName = null, LastName = "Smith" }), test.Logger))
                 .AssertException<InvalidOperationException>("First name is required.");